==> 00_data_select.ipynb <==
# Data Selection

import geopandas as gpd
import json
from shapely.geometry import box

# gda = 'data\GCCSA_2021_AUST_SHP_GDA2020\GCCSA_2021_AUST_GDA2020.shp'
# gdf_gda = gpd.read_file(gda)

# sydney = gdf_gda[gdf_gda['GCC_CODE21'] == '1GSYD']

# minx = sydney.total_bounds[0]
# miny = sydney.total_bounds[1]
# maxx = sydney.total_bounds[2]
# maxy = sydney.total_bounds[3]


# minx = 150.300861
# miny = -34.423526
# maxx = 151.589260
# maxy = -33.227966

minx = 150.301560
miny = -34.316218
maxx = 151.477327
maxy = -33.403931

crs = "EPSG:4326"

geom = box(minx, miny, maxx, maxy)

bbox = gpd.GeoDataFrame(
    geometry=[geom],
    crs=crs
)


bbox_reproj = bbox.to_crs("EPSG:3577")

bbox_reproj.to_file("bbox_sm.gpkg", driver="GPKG")

area = bbox_reproj.area[0]
bounds = bbox_reproj.bounds.loc[0]
width = bounds.maxx - bounds.minx
height = bounds.maxy - bounds.miny

print(area/ 1e6)
print(height)
print(width)
